[#2320] Make Spatial extension available for jpa-codegen
Fixes #2320
Closes #2322
