+++ querydsl-jpa-codegen/src/main/java/com/querydsl/jpa/codegen/AbstractDomainExporter.java
@@ -118,6 +118,7 @@ public abstract class AbstractDomainExporter {
         module.bind(CodegenModule.PREFIX, namePrefix);
         module.bind(CodegenModule.SUFFIX, nameSuffix);
         module.bind(CodegenModule.KEYWORDS, Constants.keywords);
+        module.loadExtensions();
         this.queryTypeFactory = module.get(QueryTypeFactory.class);
         this.typeMappings = module.get(TypeMappings.class);
         this.embeddableSerializer = module.get(EmbeddableSerializer.class);
