+++ querydsl-spatial/src/main/resources/META-INF/services/com.querydsl.codegen.Extension
@@ -1,0 +1,1 @@
+com.querydsl.spatial.apt.SpatialSupport
