+++ querydsl-sql-codegen/pom.xml
@@ -36,11 +36,6 @@
       <artifactId>querydsl-sql</artifactId>
       <version>${project.version}</version>
     </dependency>
-    <dependency>
-      <groupId>com.querydsl</groupId>
-      <artifactId>querydsl-sql-spatial</artifactId>
-      <version>${project.version}</version>
-    </dependency>        
 
     <dependency>
       <groupId>org.slf4j</groupId>
